[type]:   fix [module]: widgets [id]:     #t-75294 [desc]:   Add a flag to ensure a component may only be destroyed once. This will fix a series of issues, but also fixes the one mentioned in the ticket.
git-svn-id: http://code.extjs.com/svn/ext/trunk@4894 c0a82769-a6a8-4f8a-96ba-7530dc2e272d

+++ src/widgets/Component.js
@@ -1096,19 +1096,22 @@ var myGrid = new Ext.grid.EditorGridPanel({
      *
      */
     destroy : function(){
-        if(this.fireEvent('beforedestroy', this) !== false){
-            this.beforeDestroy();
-            if(this.rendered){
-                this.el.removeAllListeners();
-                this.el.remove();
-                if(this.actionMode == 'container' || this.removeMode == 'container'){
-                    this.container.remove();
-                }
-            }
-            this.onDestroy();
-            Ext.ComponentMgr.unregister(this);
-            this.fireEvent('destroy', this);
-            this.purgeListeners();
+        if(!this.isDestroyed){
+	        if(this.fireEvent('beforedestroy', this) !== false){
+	            this.beforeDestroy();
+	            if(this.rendered){
+	                this.el.removeAllListeners();
+	                this.el.remove();
+	                if(this.actionMode == 'container' || this.removeMode == 'container'){
+	                    this.container.remove();
+	                }
+	            }
+	            this.onDestroy();
+	            Ext.ComponentMgr.unregister(this);
+	            this.fireEvent('destroy', this);
+	            this.purgeListeners();
+	            this.isDestroyed = true;
+	        }
         }
     },
